+++ pages/index.tsx
@@ -2,25 +2,17 @@ import Form from '@components/form/Form'
 import Layout from '@components/layout/Layout'
 import TeamContext from '@context/TeamContext'
 import CookieConsent from 'react-cookie-consent'
-import { useContext } from 'react'
+import React, { useContext } from 'react'
 import style from './style.module.scss'
 import UserContext from '@context/UserContext'
 import '@reach/combobox/styles.css'
-import { TeamMember } from '@adibkhan/pogo-web-backend'
 import ImageHandler from '@common/actions/getImages'
 import classnames from 'classnames'
 import { Icon } from '@components/icon/Icon'
-import {
-  ListboxInput,
-  ListboxButton,
-  ListboxPopover,
-  ListboxList,
-  ListboxOption,
-  ListboxPopoverProps,
-} from '@reach/listbox'
 import '@reach/listbox/styles.css'
 import LanguageContext from '@context/LanguageContext'
 import { StringsType } from '@common/actions/getLanguage'
+import TeamSelector from '@components/team_selector/TeamSelector'
 
 function truncateString(str: string, num: number) {
   if (str.length > num) {
@@ -28,41 +20,6 @@ function truncateString(str: string, num: number) {
   } else {
     return str
   }
-}
-
-const right: ListboxPopoverProps['position'] = (targetRect, popoverRect) => {
-  const triggerCenter = targetRect!.left + targetRect!.width
-  const left = triggerCenter - popoverRect!.width
-  const maxLeft = window.innerWidth - popoverRect!.width - 2
-  return {
-    left: Math.min(Math.max(2, left), maxLeft) + window.scrollX,
-    top: targetRect!.bottom + 8 + window.scrollY,
-  }
-}
-
-interface TeamOptionProps {
-  team: TeamMember[]
-  name: string
-  id: string
-}
-
-const TeamOption: React.FC<TeamOptionProps> = ({ team, name, id }) => {
-  const imagesHandler = new ImageHandler()
-
-  return (
-    <ListboxOption value={id} className={style.option}>
-      <strong>{name}</strong>
-      <br />
-      {team.map((member: any, index: number) => (
-        <img
-          key={index}
-          src={imagesHandler.getMini(member.sid)}
-          className={style.member}
-          alt={member.speciesName}
-        />
-      ))}
-    </ListboxOption>
-  )
 }
 
 const myColor = '#FCAC89'
@@ -120,37 +77,7 @@ const HomePage = () => {
                   </div>
                 </div>
               </div>
-              <ListboxInput
-                aria-label={strings.select_team}
-                onChange={onSelect}
-                className={style.listbox}
-                defaultValue="defaultTeam"
-              >
-                <div className={style.changeWrap}>
-                  <ListboxButton className={style.changeTeam}>
-                    <Icon name="down" />
-                  </ListboxButton>
-                  <small>{strings.change_team}</small>
-                </div>
-                <ListboxPopover className={style.popover} position={right}>
-                  <ListboxList>
-                    {user && user.teams && user.teams.length > 0 ? (
-                      user.teams.map((x) => (
-                        <TeamOption
-                          team={x.members}
-                          name={x.name}
-                          id={x.id}
-                          key={x.id}
-                        />
-                      ))
-                    ) : (
-                      <p className={style.noteams}>
-                        {strings.team_not_available}
-                      </p>
-                    )}
-                  </ListboxList>
-                </ListboxPopover>
-              </ListboxInput>
+              <TeamSelector onSelect={onSelect} />
             </div>
             <Form />
           </section>
