+++ pages/style.module.scss
@@ -83,49 +83,22 @@
     font-size: 1.3em;
   }
 }
-
-.changeWrap {
-  display: flex;
-  align-items: center;
-  text-align: center;
-  flex-direction: row-reverse;
-  margin-left: 1em;
-  margin-right: 1em;
-  margin-top: 0.25em;
-  @include md() {
-    margin-right: 0;
-    margin-right: 2em;
-    flex-direction: column;
-  }
+.member {
+	height: 2.3em;
+	width: 2.3em;
+	object-fit: contain;
 }
-
-.changeTeam {
-  width: 4.1em;
-  height: 4em;
-  background-color: $color-brand-primary;
-  border: none;
-  box-shadow: 0.25rem 0.25rem 0.25rem rgba($color: $color-dark, $alpha: .15);
-  margin-bottom: 0.5em;
-  margin-left: 0.5em;
-  &:hover {
-    background-color: darken($color-brand-primary, 10%);
-  }
-  svg {
-    fill: $color-white;
-    height: 40px;
-    width: 40px;
-    margin-top: -14px;
-    margin-left: 3px;
-  }
-  @include md() {
-    margin-left: 0;
-  }
+.members {
+	.member {
+		@include md() {
+		height: 3em;
+		width: 3em;
+		}
+	}
+	.member + .member {
+		margin-left: 1em;
+	}
 }
-
-.noteams {
-  min-width: 200px;
-}
-
 .option {
   img {
     height: 1.5em;
@@ -139,23 +112,6 @@
   border-bottom: 1px solid $color-neutral-light;
 }
 
-.member {
-  height: 2.3em;
-  width: 2.3em;
-  object-fit: contain;
-}
-
-.members {
-  .member {
-    @include md() {
-      height: 3em;
-      width: 3em;
-    }
-  }
-  .member + .member {
-    margin-left: 1em;
-  }
-}
 
 .links {
   a {
@@ -169,7 +125,3 @@
   }
 }
 
-.popover {
-  max-height: 400px;
-  overflow-y: scroll;
-}
