+++ src/components/team_member_selector/TeamMemberSelector.tsx
@@ -438,7 +438,7 @@ const TeamMemberSelector = (props: {
 
   const getSpeciesName = (pokemon: pokemonType | TeamMember) => {
     if (!pokemon) {
-      return ""
+      return ''
     }
     if (!pokemonNames || !pokemonNames[pokemon.speciesId]) {
       return pokemon.speciesName
